pyproject.toml
[project]
name = "magic-umbrella"
version = "0.1.0"
description = "Calendar-based time allocation system using Microsoft Graph API"
readme = "README.md"
requires-python = ">=3.9"
authors = [
    { name = "Federica Nocera" }
]
dependencies = [
    "msgraph-sdk>=1.0.0",
    "azure-identity>=1.15.0",
    "msal>=1.26.0",
    "openai>=1.10.0",
    "pandas>=2.0.0",
    "python-dateutil>=2.8.0",
    "pytz>=2024.1",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "python-dotenv>=1.0.0",
    "typer>=0.9.0",
    "rich>=13.0.0",
    "pyyaml>=6.0.0",
    "thefuzz>=0.22.0",
    "python-Levenshtein>=0.25.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=4.1.0",
    "ruff>=0.2.0",
]
nlp = [
    "fuzzywuzzy>=0.18.0",
    "python-Levenshtein>=0.25.0",
]
web = [
    "fastapi>=0.109.0",
    "uvicorn>=0.27.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.ruff]
line-length = 100
target-version = "py39"

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "N",   # pep8-naming
    "UP",  # pyupgrade
]
ignore = []

[tool.ruff.format]
quote-style = "double"
indent-style = "space"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "--verbose",
    "--color=yes",
    "--strict-markers",
]
markers = [
    "integration: Integration tests requiring external services",
    "unit: Unit tests that don't require external services",
]
